+++ addons/mysterious_egg/static/src/builder/components/WeButton.xml
@@ -1,11 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-<t t-name="mysterious_egg.Button">
+<t t-name="mysterious_egg.WeButton">
     <button type="button" class="btn btn-primary"
         t-att-class="{'active': this.state.isActive}"
         t-att-title="props.title"
-        t-on-click="() => props.onClick()">
+        t-on-click="() => this.onClick()"
+        t-on-mouseenter="() => this.onMouseenter?.(props.id)"
+        t-on-mouseleave="() => this.onMouseleave?.(props.id)">
         <img t-if="props.iconImg" t-att-src="props.iconImg" t-att-alt="props.iconImgAlt"/>
         <t t-if="props.label"  t-out="props.label"/>
     </button>
