+++ addons/mysterious_egg/static/src/builder/components/defaultComponents.js
@@ -2,6 +2,7 @@ import { DropdownItem } from "@web/core/dropdown/dropdown_item";
 import { ButtonGroup } from "./ButtonGroup";
 import { Dropdown } from "@web/core/dropdown/dropdown";
 import { ToolboxRow } from "./ToolboxRow";
+import { WeButton } from "./WeButton";
 import { Button } from "./Button";
 
 export const defaultOptionComponents = {
@@ -9,5 +10,6 @@ export const defaultOptionComponents = {
     Dropdown,
     DropdownItem,
     ButtonGroup,
+    WeButton,
     Button,
 };
