+++ addons/mysterious_egg/static/src/builder/components/options/LayoutOption.xml
@@ -31,10 +31,10 @@
     <t t-else="">
         <ToolboxRow label="'└ Hor. Alignment'">
             <ButtonGroup>
-                <Button actions="{setClass: 'align-items-start'}" title="'Align Top'" iconImg="'/website/static/src/img/snippets_options/align_top.svg'" applyTo="'.row'" />
-                <Button actions="{setClass: 'align-items-center'}" title="'Align Middle'" iconImg="'/website/static/src/img/snippets_options/align_middle.svg'" applyTo="'.row'" />
-                <Button actions="{setClass: 'align-items-end'}" title="'Align Bottom'" iconImg="'/website/static/src/img/snippets_options/align_bottom.svg'" applyTo="'.row'" />
-                <Button actions="{setClass: 'align-items-stretch'}" title="'Stretch'" iconImg="'/website/static/src/img/snippets_options/align_stretch.svg'" applyTo="'.row'" />
+                <WeButton actions="{setClass: 'align-items-start'}" title="'Align Top'" iconImg="'/website/static/src/img/snippets_options/align_top.svg'" applyTo="'.row'" />
+                <WeButton actions="{setClass: 'align-items-center'}" title="'Align Middle'" iconImg="'/website/static/src/img/snippets_options/align_middle.svg'" applyTo="'.row'" />
+                <WeButton actions="{setClass: 'align-items-end'}" title="'Align Bottom'" iconImg="'/website/static/src/img/snippets_options/align_bottom.svg'" applyTo="'.row'" />
+                <WeButton actions="{setClass: 'align-items-stretch'}" title="'Stretch'" iconImg="'/website/static/src/img/snippets_options/align_stretch.svg'" applyTo="'.row'" />
             </ButtonGroup>
         </ToolboxRow>
     </t>
